Consolidate scheduler modules
